+++ backend-nest/src/english/word/word.module.ts
@@ -1,10 +1,11 @@
 import { Module } from '@nestjs/common';
 import { EnglishWordController } from './word.controller';
 import { EnglishWordService } from './word.service';
+import { EnglishWordTestService } from './test/test.service';
 
 @Module({
   imports: [],
   controllers: [EnglishWordController],
-  providers: [EnglishWordService],
+  providers: [EnglishWordService, EnglishWordTestService],
 })
 export class EnglishWordModule {}
